move hub telemetry to avoid circular dependency in cli
