+++ guardrails/validator_service/async_validator_service.py
@@ -9,7 +9,7 @@ from guardrails.classes.validation.validation_result import (
     PassResult,
     ValidationResult,
 )
-from guardrails.telemetry.hub_tracing import async_trace
+from guardrails.hub_telemetry.hub_tracing import async_trace
 from guardrails.telemetry.validator_tracing import trace_async_validator
 from guardrails.types import ValidatorMap, OnFailAction
 from guardrails.classes.validation.validator_logs import ValidatorLogs
